Fix #12. Incremented to v1.0.6. Upgraded versions of Kotling and JVM usage. Recompiled C++ projects with VS 2022.

+++ src/main/kotlin/uk/co/strimm/flowMethods/RatioFlowMethod.kt
@@ -52,7 +52,7 @@ open class RatioFlowMethod() : FlowMethod {
 
 
             for (j in 0..im.w*im.h - 1) {
-                pix[j] = (pix1[j].toDouble() / pix0[j].toDouble()).toShort()
+                pix[j] = (pix1[j].toDouble() / pix0[j].toDouble()).toInt().toShort()
             }
             statusID = 1
         }
